Additional functions to import
In an attempt to clean up the main script some functions have been
written in this file and imported. It is difficult to do this for more
elements of the project because of the way that Qt requires classes to
be used, so more or less every feature of the Gui has to be configured
within the Mainwindow class.