Adaptive GUI
The changes made for these files make it so that GUI can accept any form
of control file and it will adapt itself to accomadate thee headings and
options of that config file. Automatic loading of these has not been
done yet, so if a model other than SOL1D is to be used it has to be
manually changed in the config file.